% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{build_db}
\alias{build_db}
\title{builds a DB file, either from scratch or append to existing one}
\usage{
build_db(
  db_name,
  add_csv_name,
  verbose = TRUE,
  db_dir =
    "/Users/andrewlu/Dropbox/Documents/Labowitz/Experiments/AL-AcousticTransf/data/DNA_libraries/",
  add_dir =
    "/Users/andrewlu/Dropbox/Documents/Labowitz/Experiments/AL-AcousticTransf/data/DNA_library_additions/",
  skip_wells = FALSE
)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
