% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_wells.R
\name{check_n_wells}
\alias{check_n_wells}
\title{check how many wells per experiment}
\usage{
check_n_wells(exp_df)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
