% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_subtract.R
\name{db_subtract}
\alias{db_subtract}
\title{subtracts out volumes used during experiment and saves a new DB}
\usage{
db_subtract(
  exp_df,
  db_df,
  v_subtract_multiplier = 1.1,
  db_name,
  just_checking = FALSE
)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
