% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_biomek96w_files.R
\name{gen_biomek96w_files}
\alias{gen_biomek96w_files}
\title{generates instructions for biomek to transfection 96w plate}
\usage{
gen_biomek96w_files(
  df,
  stamp_multiplier = 1.2,
  ul_opti_to_DNA,
  v_stampplate_to_optiDNAplate,
  ul_fugene_per_microgram
)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
