% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_exp_anot.R
\name{gen_exp_anot}
\alias{gen_exp_anot}
\title{generated annotation file for downstream experiment analysis}
\usage{
gen_exp_anot(echo_df, transf_plate_type, anot_save_dir)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
