% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_source_plate.R
\name{gen_source_plate}
\alias{gen_source_plate}
\title{generates instructions for filling source plate}
\usage{
gen_source_plate(df)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
