% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_of_list_of_wells.R
\name{get_list_of_list_of_wells}
\alias{get_list_of_list_of_wells}
\title{xx}
\usage{
get_list_of_list_of_wells(pt)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
