% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plate_wells.R
\name{get_plate_wells}
\alias{get_plate_wells}
\title{gets plate wells}
\usage{
get_plate_wells(n_rows, n_cols, rows_offset = 0, cols_offset = 0)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
