% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_ng_and_db.R
\name{join_ng_and_db}
\alias{join_ng_and_db}
\title{joins the ng and db DFs, and checks / resolves any conflicts}
\usage{
join_ng_and_db(ng, db)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
