% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_exp.R
\name{new_exp}
\alias{new_exp}
\title{makes a list of experiment parameters}
\usage{
new_exp(name, ng_csv_name, db_csv_name, n_replicates, n_cell_types)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
