% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_n_wells.R
\name{print_n_wells}
\alias{print_n_wells}
\title{check how many wells per experiment}
\usage{
print_n_wells(df)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
