% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ng_input.R
\name{process_ng_input}
\alias{process_ng_input}
\title{processes ng input file (wide to long, clean up..etc)}
\usage{
process_ng_input(experiment_name, ng_dir, n_replicates)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
