% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake_ng_csv.R
\name{remake_ng_csv}
\alias{remake_ng_csv}
\title{remake ng csv file from script outputs}
\usage{
remake_ng_csv()
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
