% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_checks.R
\name{run_checks}
\alias{run_checks}
\title{runs checks}
\usage{
run_checks(all_experiments, check_db, save_plasmids_needed_csv)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
