% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAcousticTransf.R
\name{runAcousticTransf}
\alias{runAcousticTransf}
\title{main script that runs everything}
\usage{
runAcousticTransf(all_experiments)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
