% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_write_csv.R
\name{safe_write_csv}
\alias{safe_write_csv}
\title{safely writes csv (doesn't overwrite existing files)}
\usage{
safe_write_csv(df, dir_save, add_date_time = FALSE, verbose = FALSE)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
