% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{set_up}
\alias{set_up}
\title{sets up experiment variables}
\usage{
set_up(
  runs_dir =
    "/Users/andrewlu/Dropbox/Documents/Labowitz/Experiments/AL-AcousticTransf/runs/",
  run_name,
  db_dir =
    "/Users/andrewlu/Dropbox/Documents/Labowitz/Experiments/AL-AcousticTransf/data/DNA_libraries/",
  transf_plate_type,
  write_files
)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
