% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{sets up experiment variables}
\usage{
setup(
  home_dir,
  run_name,
  ng_csv,
  n_reps,
  n_celltypes,
  source_plate_max_v,
  source_plate_dv,
  dna_echo_conc,
  dest_plate_type
)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
