% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_echo_csv.R
\name{gen_echo_csv}
\alias{gen_echo_csv}
\title{given db and experiment dataframes, generate echo instructions csv}
\usage{
gen_echo_csv(
  exp_df,
  db_df,
  transf_plate_type,
  echo_csv_save_dir,
  info_csv_save_dir
)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
