% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_acoustic_transf.R
\name{run_acoustic_transf}
\alias{run_acoustic_transf}
\title{main script that runs everything}
\usage{
run_acoustic_transf(all_experiments, gen_anot = TRUE)
}
\arguments{
\item{xxxx}{param}
}
\value{
xxxx return
}
\description{
This function xxxxxxx
}
